Reorganize JarResultBuildStep and split it
This class had become completely unmanageable due to its size.
Given I'm willing to invest some time to see if we can improve our zip
build time, this is a necessary step in preparation to the upcoming
improvements.
