+++ core/deployment/src/main/java/io/quarkus/deployment/pkg/jar/LegacyThinJarFormat.java
@@ -1,0 +1,6 @@
+package io.quarkus.deployment.pkg.jar;
+
+public class LegacyThinJarFormat {
+
+    public static final String LIB = "lib";
+}
